Extracted zone distribution feature into a separated source file
